BI-192 simplify type checking and add date arithmetic

+++ evaluator/test_setup.go
@@ -258,7 +258,7 @@ schema:
      -
         Name: a
         MongoType: int
-        SqlType: numeric
+        SqlType: int
      -
         Name: b
         MongoType: int
